updated windows cmd file that copies windows release files to a release archive.
(work in progress but this one actually works - it is the one I use to get me files to be used by my VS2012 c++ DLL). Would like to have it in the release branch as it is part of the procedure to create a directory with VS2012 lib and header files to be used to create the c++ DLL part of a C# release.