Add another operator= overload for BasicRow
When T == U, the compiler-generated default `operator=` is used rather than the
`template<class U> BasicRow& operator=(BasicRow<U>)` overload, which meant that
assigning a Row to a ConstRow or vice-versa worked correctly, but assigning a
Row to a Row did not.

None of the current tests actually verified that the user-supplied assignment
operator was being used, so I expanded the
Table_RowAccessorAssignMultipleTables test to also test assigning a Row to
another Row (in addition to assigning a RowExpr to a Row).