Fix unbounded growth of the write log when committing from a single SharedGroup
Entries in the write log were only ever released when advance_read actually
changed the version, but when every write is made from a single SharedGroup,
that group's version is always the latest. As a result, the WriteLogRegistry
did not get notified that the writing SG did not need the old entries until
that SG is destructed.

Fix this by telling the TransactLogRegistry to release older entries in
advance_read even if the version didn't change. It'd be nicer to call
release_commit_entries in commit_and_continue_as_read, but that function
currently doesn't have access to the TransactLogRegistry.