Workaround to load libcrypto.so for android
Fix #1006
On some korean version LG G3 phone, a so lib named libsktdrm_sf.so
exists in the /system/lib and contains the same crypto functions we need
which are supposed to be in libcrypto.so. This caused some ramdom
issues.

To avoid that, we always try to load from libcrypto.so on Android first,
if there isn't, try search in other system libs.

And also, if the function doesn't exists, a crash should happen in the
release build to make debugging easier.