Embed crypto functions from openssl for android
Revert dc33400 with modification
Fix #1006
On some korean version LG G3 phone, a so lib named libsktdrm_sf.so
exists in the /system/lib and contains the same crypto functions we need
which are supposed to be in libcrypto.so. This caused some random
issues.

Also, crypto related functions have been seen in other so files under
/system/lib for many devices.

To avoid all of those random issues, we just embed those functions in
our lib.