Added copy constructor for BasicRow
There was already a copy constructor defined for BasicRow, but with the
following signature:

    template<class T> template<class U>
    BasicRow<T>(const BasicRow<U>&)

Because the argument has a different template argument (U instead of
T), the default copy constructor was being used. This could cause an
issue as demonstrated in the attached unit test, or when compiling
with the `-fno-elide-constructors` compiler flag.

The issue caused such copy constructed row accessors to not be properly
registered with the table, and wouldn't get updated upon changes.