Remove an outdated comment about Valgrind.
Prior to 2136f4bb, `Value` didn't track how many items were in use so it
was necesary to initialize all items to prevent warnings about the use
of uninitialized memory. `Value` now tracks how many items are in use
and the code the comment is referring to has been removed.