Fix Query_DeepCopy with `REALM_OLDQUERY_FALLBACK` disabled.
In 5ad0e268, `Query` moved to caching its current descriptor rather than
computing it on demand. This brings with it a requirement that the
descriptor be refetched when the `Query`s table changes. This was not
being done for the two `Expression` subclasses (`UnaryLinkCompare` and
`Compare`), resulting in a null pointer dereference when running the
tests with `REALM_OLDQUERY_FALLBACK` disabled.

This would have been caught by tests, but Query_DeepCopy's attempts to
test the query expression codepaths with `REALM_OLDQUERY_FALLBACK`
enabled were not successful.