Ensure that the `constexpr` parts of `Optional` compile when used in constant expressions.
In order to be used in constant expressions, `Optional<T>` must be a
literal type. One constraint on such types is that they are trivially
destructible (i.e., their destructor must do nothing). `Optional<T>` can
satisfy this constraint when `T` is itself trivially destructible. To
accomplish this, we split `Optional`s storage into a separate
`OptionalStorage` type that is specialized based on whether `T` is
trivially destructible. In the trivially-destructible case, the default
destructor is used. In the non-trivially-destructible case, the value's
destructor is explicitly called when the value is engaged.

A second constraint on literal types is that their `constexpr` constructors
must initialize all data members. To achieve this without needing to
initialize the storage used by `T` when constructing a disengaged
`Optional`, we store `T` in an anonymous union along with an empty type.
The empty type is initialized when constructing a disengaged
`OptionalStorage`, with the `T` member being initialized in other cases.
An extra advantage that this technique provides is that the compiler
will track which member of a union in a literal type is active, and will
provide a diagnostic if the program attempts to access any other
member. The end result is that a diagnostic will be generated if a user
attempts to dereference a disengaged `constexpr Optional<T>`.