Allow comparisons involving unary links on both sides of the operator.
The comparison code currently behaves as though links indicate "many-to-x"
relationships, but that's only the case when dealing with a link
list. Unary links are in fact "one-to-x" relationships, much like normal
column values.

This change results in `Columns<T>::evaluate` producing a `Value` with
`from_link_list` set to `false` when only unary links are involved. This
has the desired effect of sending comparisons and operators down the
"one-to-x" paths rather than the "many-to-x" paths. This prevents
comparisons with unary links on both sides from asserting due to the
"many-to-many" case being undefined.

A side-effect of this change is in how `Columns<T>::evaluate` handles
null unary links. Formerly, an empty `Value` was produced. Now a `Value`
containing a single null is produced.