Remove `QueryGroup::update`, `QueryGroup::update_override`, and `Query::subtables`.
This leaves only a single pointer to each query node within the group,
which should allow us to switch the nodes over to explicit unique
ownership.

The purpose that `update` was serving for `Not` and `subtable` is
replaced by moving reponsibility for creating the node to
`handle_pending_not` and `end_subtable` respectively. The body of their
group is reparented as the `NotNode` / `SubtableNode`'s condition.

`OrNode`'s use of `update_override` requires special handling.
`update_override` was used to have the next node that was added be
appended to `OrNode::m_condition`, and for future nodes to be added as
children of the most recently added condition. We make this sequence
explicit with the `QueryGroup::State` enumeration, which influences the
behavior of `Query::add_node`. When `Query::Or` is called, the current
group's state is set to `OrCondition`. When the next node is added, the
state transitions to `OrConditionChildren`.