Add an overload of `Query::and_query` that takes an rvalue.
This avoids copying the node hierarchy in the extremely common case that we're
combining the query with a temporary.

In the future it may be beneficial for `Query::operator&&` and `Query::operator||`
to be overloaded for rvalues as well. I'm not adding those overloads at present as
the methods are not used by the Cocoa or Android bindings, so it doesn't seem worth
the complexity that the additional overloads per method would add.