Use `std::string` rather than introducing `OwnedStringData`.
`StringNodeBase` now stores `m_value` as an `Optional<std::string>`. To make this ergonomic,
`StringData` gains a converting constructor from `util::Optional<std::string>` to propagate
nullability from the `Optional`.

`StringNode` now stores `m_ucase` and `m_lcase` as `std::string`s. To avoid unnecesary heap
allocations during their initialization, the signature of `case_cmp` is updated to return an
`Optional<std::string>`. This allows the small string optimization to store data on the stack
before we move it into `StringNode`s members. The disengaged state of the `Optional` is used
to communicate a failure to convert the input to the desired case. Since a number of callers
of `case_map` don't deal with errors from the function, an overload is added with a flag to
suppress errors. In this case the return value in the error case is simply an empty string.