Update to SSE/AVX detection on gcc/clang
Instead of using assembly, we can use intrinsics created by GCC and
adopted by clang in order to do run-time detection of the specific
extensions provided by the CPU. This approach is a lot more portable
and maintainable.

Support for `__builtin_cpu_supports()` was introduced in GCC 4.8[1] and
clang (at least 3.7). A full list of detectable extensions is available
here[2].

[1]: https://gcc.gnu.org/gcc-4.8/changes.html
[2]: https://gcc.gnu.org/onlinedocs/gcc-4.9.2/gcc/X86-Built-in-Functions.html