Make all feature macros always defined
Having the feature macros always defined allows us to us to do the
following:

    #if FOO

instead of

    #ifdef FOO

The main issue with `#ifdef` is that it doesn't catch typos, and can
thus sometimes lead to hard to detect issues. Issues like this are also
very hard to detect because you can't easily grep for the typo. Because
all defines macros resolve down to bools or 0/1 ints, they can also be
used directly in CPP code, without the need to wrap it them in
preprocessor logic (`#if`):

    if (REALM_PLATFORM_APPLE_TVOS) {
        // ...
    }
    else if (REALM_PLATFORM_APPLE_IOS) {
        // ...
    }

Obviously, preprocessor conditionals are still needed to prevent
unresolved symbols, etc.