Fix `Optional`s buggy copy and move assignment operators.
Both assignment operators failed to set `m_engaged` when assigning an
engaged optional to an unengaged optional. The lack of test coverage for
these operators caused this issue to be missed.

Additionally, `Optional`s value assignment operator was insufficiently
constrained resulting in compile errors when attempting to use the copy
assignment operator due to ambiguity between the copy assignment and
value assignment operators. To address this, the value assignment
operator is now constrained to argument types as per
[optional.object.assign.11][1]

[1]: http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2014/n3848.html#optional.object.assign