Use universal references in range-based for loops
The reason to prefer universal references in range-based for loops is
that dereferencing itertors in C++11 doesn't necessarily yield an
lvalue. Thus, using `auto&` or `const auto&` would not be able to
capture that.

The rest of this PR follows the same style, and only uses const
references when it is explicitly declared as such.

Alternatively, it would be possible to have

    template <typename T> T const& as_const(T const& v) { return v; }

which would then give us the ability to use

    for (auto&& x : f) { ... } // non-necessarily const
    for (auto&& x : as_const(f)) { ... } // always const

and be even more explicit.