Set the last seen version correctly for TableViews created from queries on LinkViews
The initial last seen version was always being set to the target's version,
while is_in_sync() uses the origin table's version. This normally just resulted
in the query being rerun when it didn't need to be, but could theoretically
result in it not being rerun when needed given sufficiently bad luck.