Implement support for subqueries.
This provides subqueries using the syntax from #905. The evaluation of
subqueries operates by explicitly evaluating the subquery for each
linked row and counting the number of linked rows that matched. In the
future we can short-circuit evaluation when the target match count is
exceeded for subqueries of the form:

    table1->column<LinkList>(col_link2, table2->column<Int>(0) == 1).count() > 0

Fixes #905.