Throw exceptions on decryption failure rather than terminate()
This will still normally result in terminate() being called due to throwing out
of a noexcept function, but it makes it so that the detection of invalid
encryption keys on open is not reliant on the first page failing to decrypt and
simplifies things.

The actual bug fixed by this is that when the first four bytes of the file were
zero the first page would be treated as uninitialized rather than encrypted
with a different key, and then reading the footer would call terminate() if the
IV for that file position was not also zero.