Use SDK identifiers rather than walking the file system to discover available SDKs.
The presence of SDKs is confirmed using `xcodebuild -version -sdk foo`,
and the compiler is now invoked using `xcrun -sdk foo c++`.

This changes the file names of the iOS, tvOS and watchOS libraries built
by `build-cocoa` as they now include the SDK identifier rather than a
title-case variant of the platform name.