Adjust the fix as suggested by @kspangsege.
`avail` is changed to `version_type` rather than encoding the assumption
that `version_type` is `uint_fast64_t`, `n` is reverted to be `size_t`,
and a cast is added when assigning `avail` to `n`. The cast is safe due
to the if statement that the assignment is performed within.