Add width-specialized version of Array::adjust_ge()
This is called a whole bunch of times in advance_read() when there's a live
TableView for a table that had rows inserted or deleted. With a TableView with
100k rows and a transaction with 33k insertions and 33k deletions this cuts the
time spent in adj_row_acc_insert_rows() from 9 seconds to 2.5 seconds and the
total runtime of advance_read() from 11 seconds to 4 seconds.