Convert this to a test-case and use test-suite
Doing this allows to continue using the benefits of SHARED_GROUP_TEST_PATH,
i.e. that the path is deleted when it goes out of scope. Since the path is
deleted when it goes out of scope, there is no need to waste energy crafting a
nice filename for it.