Have `Table`'s reference counting use atomic operations.
In #826, it was made safe to destroy a `Row` accessor from any thread as
it's difficult to prevent users of the Cocoa binding from doing this
unintentionally. An oversight in this change is that `Row` holds a
reference to the `Table` it belongs to, but `Table`s reference counting
is not threadsafe. This means there is the potential for adjustments to
the reference count to be missed due to concurrent adjustments on
another thread. We address this by switching to atomic operations for
manipulating the reference count.

Note that this change does not make it safe for the final reference to
a `Table` to be dropped from an arbitrary thread. `Table`s destructor
interacts with other objects, so correct synchronization is necessary to
ensure that no data races occur during the destructor's execution.