Improve the `CHECK_TABLE_VIEW` macro.
Move most of the logic into a helper function. This avoids issues with
the use of local variables in the macro conflicting with those in the
caller's scope. It also makes it easier to improve the output by using
the expressions passed to `CHECK_TABLE_VIEW` in the output in the
failure cases.

(cherry picked from commit 7a14a6a46bea819d198b3f44ff632dc3a77bce7d)