Don't throw an exception from make_dir() when the directory already exists
And instead return a bool for that.

Checking if the directory already exists before trying to create it runs into
TOCTOU issues, so the directory already existing can't reasonably be considered
an error case. Swallowing the exception when the calling code doesn't care
about it makes debugging more difficult, and when it does need to do something
different, using the exception for control flow would be ugly.