Lock the write log header file rather than creating a separate lockfile
Unlike with the main lockfile, nothing else is locking WriteLogCollector's
header file, so we can use that as the file for the InterprocessMutex rather
than creating a new one.

This fixes a performance regression introduced in #1446, which more than
doubled the time required to commit small transactions due to all of the extra
`open()` calls introduced. No release notes entry since #1446 has not yet been
included in a release.