Have `TableView` and `Query` expose whether their rows are guaranteed to be in table order.
The change calculation logic in Cocoa's fine-grained notifications has
an optimization for rows guaranteed to be in table order, but knowing
when this optimization can be activated requires tracking the source of
each `TableView`. This is unnecessary given that a `TableView` itself
already knows its source and therefore knows whether its rows are
guaranteed to be in table order.