Fix incorrect comparison semantics when comparing a LinkList column with a Row using a query expression.
The old query fallback path that attempts to take advantage of
`Query::links_to`'s performance was being selected in cases where it has
different semantics than we need. In particular, the sequence of
`Query::Not` / `Query::links_to` on a link list column only matches rows
where none of the links is equal to the row we're looking
for. Comparisons against values from link lists are expected to match if
any of the values fulfills the criteria.

This was noticed due to "failing" tests when run with
`REALM_OLDQUERY_FALLBACK` disabled. The failing results were in fact the
correct results, with the test having the wrong expectations.