Fix handing over an out-of-date TableView representing backlinks to a deleted row.
Handed-over `TableView`s were using a magic version number of -1 to represent
that the view is out of sync. `TableView::outside_version` was using
this same magic version number to indicate that the view depends on a
deleted link list or row. This could lead to handing-over an out of sync
`TableView` creating a `TableView` that always believed itself to be in
sync, and thus could never be brought back in sync.

To address this we change `TableView::m_last_seen_version` to be
`Optional`, with the empty optional representing the fact that no
version has been seen yet.