Fix queries over multiple levels of backlinks to work when the tables involved have their backlink columns at different indices.
The logic for finding the backlink columns was always using the query's
root table to find the backlink column's index. That happened to work in
some cases, but died with an assertion failure in others. We now use
the current target of the link chain to find the index of the backlink
column.