Remove unnecessary fsync()s of commit log files
The writelogs do not need to survive power loss (or even ever actually hit disk
at all), so there's no need to call `File::sync()` on them. These were
incorrectly added by the combination of
0d2fd43a08734fecbaa2ceb30c7ecdd6a5bd3869 and
185de19a140c4c9cd2f6dcb46679e16db853df05: the first made all calls to
`File::resize()` call `fsync()` even though only resizes of the Realm file
itself need it, and the second fixed this by pulling the call to `fsync()` out
of `resize()`, but did so even in the places where the sync was not required.

There is a remaining call to `msync()` on the log files before flipping the
active file bit in the header which has a much smaller impact on performance.

This cuts the time to open a SharedGroup on an iPad Air from 100ms to 30ms (and
makes it not bound on disk write speed) and makes commits a bit faster (~10%
for very small ones).