Introduce REALM_NEVER and REALM_ALWAYS macros
These macros are introduced in order to be able to indicate that some
conditions are supposed to be either always or never true during unit
testing. When measuring the test coverage, the macros will defined as
'false' and 'true' resp. This will give the compiler the possibility
to optimize out the condition and in this way give a more useful coverage
score.