Adding .clang-format configuration file
This is just a proposal on how the formatting can be specified.
It works reasonably well, but it of course has some shortcomings
that we may have to live with.

To reformat a file, run:
  clang-format -i foobar.cpp

To reformat the part of the code that you have modified, run
in project root:
  git diff -U0 HEAD^ | clang-format-diff -i -p1

Has been tested with version 3.8