Move files to the correct target in the Xcode project
A bunch of the tests for util functionality were in the test-utils library
(i.e. the helper library for the tests) rather than the unit tests target,
which resulted in it not compiling. Fixing this resulted in Xcode changing the
UID for a bunch of files. I'm not sure why.