Rework Query initialization to support adding columns
This changes all query nodes to use the following pattern:

  1. On initial creation, store the passed-in column index.
  2. When the node is added to a Query, bind to that Query's table immediately
     rather than waiting until the Query is run, and never refer to the column
     index again after this.
  3. When cloning for handover, construct the new nodes using the column
     indexes from the actual columns rather than the originally passed-in values.

The third part requires making it so that columns which are part of tables keep
track of their logical index within the table. Table and Link columns did this
already (although Link columns did not properly update it after new columns
were added), so that functionality was just moved to ColumnBase.