Workaround for mkfifo bug on BB devices (#1985)
Fix for https://github.com/realm/realm-java/issues/3178

BB 10's mkfifo() has a bug. When the fifo file exists, from the
document, errno EEXIST should be set. But instead, it fails with
error ENOSYS which is translated to exception message "Function
not implemented".

To solve this, we use stat to check if the path exists and it is
a fifo.