Use shared_ptr for Descriptor
To ensure the thread safety of destructors.

Root descriptor is created by table which also keeps a shared ref
on it. Sub descriptor is created by root descriptor which also keeps
a shared ref on it. This means for most cases, the descriptors live
as long as the root table. Don't hold a shared ref on it to avoid
circular refs.