Remove definition of TableView destructor
Having the destructor defined, prevents the generation of at least the
move constructor and move assignment operator. This is not a problem
in released code, as the need for those functions are most probably
optimized out. In debug mode, however, we have a fundamentally different
behavior, where the copy constructors are used where we would expect move
constructor. Also the comments hints that the class is moveable.