Insert an exit in openssl/config
When config fails, the build will continue silently and use the previous
target results which is wrong.

We always have a problem that the openssl for armeabi-v7a, it fails
during the config stage. But instead of stop building, it will just use
the armeabi library which was just built previously.

Investigation of openssl armeabi-v7a build failure is needed.
Simply change the checking in openssl/config from armv[7-9]*-*-android
to arm-v[7-9]*-*-android could make it compile, but failure happens in
the final link stage.