Updating unit test StringIndex_Fuzzy
This is an updated version of the unit test that was added in dc566f5b3.
The error that this test exposed was due to the way the test was
implemented. It inserted a value into the column which was the result of
a get from the same column. As we use StringData to hold the data to be
inserted, we have a direct pointer to where the string data is stored. In
this case we use long strings which means that many strings are stored in
the same blob. If the value to be replaced is stored in the same blob as
the new value and the old value comes first, then the new value will move
in memory before the copy is done because we need to adjust for the new
size. So the pointer we get passed in becomes stale and we get the wrong
value copied in (which will have no index).

The new test case is modified in the way that we first copy the content out
in some temporary strings before inserting them again.