Fix a BadVersion exception when async notifiers run during delivery
The version for reused handover objects was being bumped on delivery to
the most recent version that had been calculated, which may be more
recent than the version actually being delivered. Instead bump it t the
point where we would normally recreate the handover object so that it is
guaranteed to get the correct version.