Fix a crash when reimporting a Query into the same SharedGroup as it was exported from
`m_table` needs to not be set on the handover Query in order for the column
references to be updated properly when re-importing it into the same
SharedGroup. This could theoretically cause problems when exporting and
importing into a different SharedGroup, but it's very unlikely (it'd require
that the source SharedGroup be destroyed and then the target one happen to
allocate the equivalent Table at the same memory address).

The removed assertion was not actually asserting anything sensible (`m_table`
was previously the same table as it was getting the column index from, so it
wasn't an interesting assert, and now is null so the assert would just crash).