Fix a crash when calling `SyncManager::reset_for_testing`.
`SyncClient`'s lifetime was managed by a `shared_ptr`, resulting in
`SyncManager::reset_for_testing` failing to destroy it when
expected. This meant that its runloop was still running, and that it
could continue to deliver callbacks to sessions that `reset_for_testing`
was in the process of destroying.

To address this `SyncClient` is now owned by `SyncManager`, with
`SyncSession`'s holding only a reference to their client. This is safe
because sync asserts that all sessions have been destroyed when the
client is destroyed.