Build the Cocoa package with -fvisibility-inlines-hidden (#2350)
Mostly just to bring it in line with Xcode's default settings, as
otherwise it produces warnings when building a dynamic framework using
the core library (although it's unclear as to why it's only recently
started to produce those warnings). It does reduce the compiled size by
a not very interesting amount (~30 KB).