Support for list of primitives
A column that should hold list of primitives is created as a sub-table
column where the sub-table has only one column matching the primitive type.
A new function - Table::add_column_list - is added for the purpose of
creating such a column. Some helper functions that will allow you to set and
get std::vector<T> variables are also supplied.

The unit test shows how to use the feature.

Issue #1881