Fix a race condition leading to bad version errors in the notifier worker
We need to decide which version to advance to in the notifier worker while the
notifier lock is held, as otherwise we may advance to a version that some
notifiers needed to skip over and then trying to perform that skip will fail.

The problematic sequence of events was:

  1. Notifier worker releases notifier lock
  2. Thread with notifications acquires notifier lock
  3. Thread with notifications commits write transaction
  4. Thread with notifications sets skip version
  5. Notifier worker advances to latest version
  6. Notifier worker acquires notifier lock
  7. Notifier worker tries to advance to skip version and fails

No new test for this as I don't know how to produce it other than hammering
commits and hoping to hit the bad timing.