Support for list-of-primitives
A column that should hold list of primitives is created as a sub-table
column where the sub-tables have only one column. A new function -
Table::add_column_list - is added for the purpose of creating such a column.

You can get an accessor object to a specific list element by calling the
Table::get_list<T> function. The object is of class List<T>, where
T is the type of the elements in the list. You access and manipulate the
list through this object.

Change-Id: Ief55af1c49451464d4fca3ce6c2aeaf5bf0c6e65