Get and set sys_tmp_dir
When compacting the Realm on external storage for Android, SharedGroup
always uses the default sys_tmp_dir which is invalid for Android. Thus
we need an API for setting the default sys_tmp_dir for bindings.
The getters is also required since ObjectStore may use it to create
pipe/fifo.

To solve java issue https://github.com/realm/realm-java/issues/4140
Part of #2411