Remove an unused code path in `TableView::do_sync`.
`Table::find_all` has delegated to `Query` for some time now,
so the FIXME in question has already been addressed. The only other way
for this code to be hit was on a `TableView` created via
`Table::get_range_view`, and that was changed to delegate to `Query` in
an earlier commit.