Rework the ownership of `SyncSession` so that it's always owned by a `shared_ptr`.
Ownership of a `SyncSession` was previously managed by a `shared_ptr`
for as long as there were external references to it. When all external
references were dropped and the `SyncSession` became inactive, it was
transferred to a `unique_ptr` held by the `SyncManager`. This transfer
proved problematic as it meant that the `weak_ptr`s that `SyncSession`
used internally to refer to itself from callbacks expired, making it
impossible to correctly handle the callbacks.

The new approach is for `SyncSession` to always be managed by a
`shared_ptr`. To detect when external references to the `SyncSession`
has gone away, we tie the lifetime of the external references to a
separate object and use the destruction of this object to signal that
the `SyncSession` has become inactive. Since the `SyncSession` is always
managed by a `shared_ptr` any `weak_ptr`s continue to work as expected
for the complete duration of the session's lifecycle.

This has the secondary benefit of removing the need to track active and
inactive sessions separately in `SyncManager`.