Make Table::m_spec a pointer into the spec/subspec structure
We must follow the rule that we cannot have multiple accessor objects
for the same file object. For this reason a subtable accessor cannot
have its own spec accessor object. Rather should all accessors to
objects in the same subtable column share the same Spec accessor object.

So the Spec is changed from being embedded in the Table class to being
referenced. The Spec objects are being managed in a tree with root in
the root table Spec object. A new class - SpecPtr - is introduced to
handle the ownership of the Spec object. Only if the table is a table
with independent spec, the spec object should be deleted when the Table
object is deleted.