Add a templated version of Table::set()
This follows the approach taken by Table::get() in making the templated
function have the actual implementation and turning the existing
non-template version into a wrapper. It also moves these wrappers to
table.hpp as they'll trivially inline out of existence.

There are no changes to the bodies of the functions here other than
moving where in the file they are.