Some link's field changes are not reported (#432)
This is reported by https://github.com/realm/realm-java/issues/4474
The root cause is the nested links is not handled well. In some cases
those indices have been saved to the m_not_modified accidentally.

The key detail is already_checking exiting early due to hitting a
circular link needs to be handled the same as exiting early due
to hitting the max recursion depth, but we still want to mark
intermediate links as unmodified if we actually did fully walk
the tree from there.