Avoid invalidating object schema pointers when the schema has not changed.
The Cocoa binding stores pointers to `ObjectSchema` instances owned by
the `Schema` that the `Realm` instance holds. Replacing the `Schema`
invalidates those pointers. To work around the specific issue seen
in the Realm Browser we can avoid replacing the schema when it has not
changed, which happens whenever new data is written by an external
process, whether sync or a different app, while the Browser has a given
Realm open. Solving this problem in the face of schema changes will
require broader changes.