Windows encryption
All places in the encryption-layer where int was used as *nix file descriptor,
this has been modified to use the new FileDesc typedef which is either int or
void* on *nix/win respectively.

Many methods on the File class have been split into static methods that take
a FileDesc as argument and do all the work, and with a thin wrapper for
non-static methods. This lets encryption utilize the File class (and save
tons of redundant code).

The native Windows library bcrypt is used for AES encryption on win32/uwp.
Kalven-SHA is used for SHA224 hash because 224 is not available on Windows.