Don't attempt to make non-additive schema changes to a synced Realm.
As of sync v2.0, object store's `metadata` table is no longer being
synced. This means that it's possible for a Realm to be downloaded and
opened and contain the expected schema, even though it has no schema
version set. If the Realm's schema doesn't match the schema specified by
the app, this could lead to us performing non-additive schema changes,
such as removing properties, in an effort to make the "new" Realm file
have the schema we expect.

We address this by always using `apply_additive_changes` when updating
the schema of a Realm using additive schema mode, rather than using
`create_initial_tables` when the Realm has no schema version. This
avoids ever attempting non-additive changes when in additive schema mode.

In addition, sync v2.0 no longer syncs the presence of indexes. This
means that we can now ensure that indexes match the current app's
desires without risking fighting with another client that disagrees
about the presence of indexes.