Add support for StringData in clusters
A new class - ArrayString - is introduced. It will handle the switch
between the 3 different storage models - short, long and blob.

It requires a new type in ColumnTypeTraits - cluster_leaf_type -
as the leaf types are different in the legacy bptree and in the clusters.
At least for StringData and BinaryData.

Change-Id: I32b77b9a4cf181c25efdb3b20c910d5cdf0c5b61