UWP improvements (#2849)
* use std::filesystem to get temp directory if available

This fixes the behavior on Windows which doesn’t have TMPDIR, and also supports UWP.

When running tests on CI we redirect TMP to a temporary folder inside the workspace which we then delete, because we risk running out of unique file names for GetTempFileName() in subsequent builds on the same machine.

* Define windows platforms in features.h

Define REALM_WINDOWS to denote regular Windows, and REALM_UWP to denote Windows Store apps. REALM_UWP is defined entirely in terms of what the standard compiler invocation has, instead of relying on CMake or someone else to provide definitions.

This makes it possible to build realm-sync for UWP without teaching it about core’s REALM_UWP.

* Enable make_temp_dir on UWP

* Use LCMapStringEx instead of CharUpper/CharLower

It’s also supported on UWP, and can be made explicitly locale-aware.

* use winsock’s declaration of timeval

* enable test on UWP

* reimplement winfork with CreateProcess

This allows tests using it to be concurrent. The bonus is that it doesn’t drag the Shell API in.

* implement more reliable timers on Windows

* address review comments

* fix wrong condition