Avoid accessing `AdminRealmListener` after it is deallocated.
If the timing is right, it's possible for the
`wait_for_download_completion` callback to be enqueued with a successful
result but have the `AdminRealmListener` destroyed before it is
invoked. If this happens, references to member variables within the
download callback would access freed memory.

To avoid this, we change `AdminRealmListener` to shared ownership so we
can use a weak pointer within the callbacks to detect whether the
`AdminRealmListener` is still alive, and to ensure that it remains alive
for the duration of the callback.