Add ListIterator class
The ListIterator contains a caching element that makes it possible
to implement the standard iterator interface where operator->() and
operator*() returns a pointer and a reference resp. As the functions
are inlined, the compiler is able to remove overhead.

Rename variables in Table_ListOfPrimitives. It is confusing that
the vectors are called '*_list' and the lists are called '*_vec'.

Change-Id: I3f7da2ae87a8dfb38bd7e4f1eeaa69115cbb619f