Add the ability to override the server that's used by a given `SyncSession`
This is for internal use by ROS as part of its support for failover. If
the `SyncSession` is active, we ask the underlying `sync::Session` to
disconnect and reconnect to the new server. We also stash away the
overridden server's details so it can be used again we later disconnect
and reconnect. If the session is not yet active, we simply stash away
the details to be used when establishing the connection.