Make queries immune to column moves
The names of the columns are stored in the query and the column id's
are then looked up each time the query is run.

One consequence of this is that the backlink columns must have names
too.

Another consequence is that columns must all have names and that they
must be unique within the table.

Change-Id: I76c735f033123ed3cd2d160187e7cb9dc6d3e495