Remove ParentNode::match_local again
It was not possible to make it work on Windows. The solution we have here
is also better in the way that it uses already existing functionality.

Some more adjustments were made in order to make the test pass on Windows.
Especially the move contructor on TableView was not working, but this was
not used on Linux as we have copy elision here.

And then some windows header defines "small" as something.

Change-Id: I36517709e8944fbeda56bb71ffe679ce3461a253