Make ListIterator stable against deletions
The list object will keep a record of which elements that has been
deleted since the creation of the list. Iterators will iterate through
the indices as they were when the list object was created, but they
will skip over the deleted indices when advanced. When the iterators
are dereferenced they will be adjusted according to the list of deleted
elements.

This is more memory efficient than having a shadow array of indices.

Change-Id: I6f20ff16c428238d51d12225f6a68657abdcff48