Support stable iteration through table
The iterator is now based on keys instead of indexes. This makes it
also faster to find next leaf when a cluster boundary is reached.

When the iterator is dereferences and the version is not up-to-date
we try to load the same key again. If the same key is found all is
well. Otherwise the key must have been deleted and the call will
throw.

If we call operator++ and the object is deleted, we will just jump to
the next object.

Change-Id: I62e77e260417025e329dba44ad79207a5def6a3d