Replace ColumnAttr with ColumnAttrMask in Spec interface
What we get and set through the Spec interface is a bit-mask that
is the result of or'ing one or more ColumnAttr values. The
mask itself is not of ColumnAttr type as it can take on many
more values than this type indicates.

ColumnAttrMask class ensures that we can only set and reset proper
bits.

Change-Id: I0439fe497922a7c053c1ca861e1ae4d4a187ab94