Support recursive delete in clusters
We will drop the old approach where the backlinks are removed
first and the objects are deleted in one go in the end.

In the new approach we delete the primary object, and record
which objects should be deleted as a consequence. Then
the process is repeated with these objects as primaries.

Change-Id: I8d74507430f94cb521bf35bd8530695e2be82cfc